Replace use of assets to getAccountAsset ethereumUtil
This particular change also required an update to a gas hook.
The gas hook from TransactionConfirmationScreen would include an asset override for network native assets, which is already handled in the redux gas.

I left a TODO JIN for consdering whether the native asset address should be used vs the uniqueId.